search filter functionality, prep for evaluation model

+++ server/utils/updateUpcomingMission.js
@@ -6,13 +6,17 @@ module.exports = {
         
         if (contact.nextMission){
             const missionCleanup = await Mission.findById(contact.nextMission)
-            if (missionCleanup.missionStatus=== "new"){
+            if (missionCleanup && missionCleanup.missionStatus=== "new"){
                 await Mission.deleteOne({_id: contact.nextMission})
             console.log("Deleted old mission")
             }
             
         }
-
-        await Contact.findByIdAndUpdate(contact._id, {nextMission: newMission._id}, {new: true})
+        try {
+            await Contact.findByIdAndUpdate(contact._id, {nextMission: newMission._id}, {new: true})
+        } catch (err) {
+            console.error(err);
+        }
+        
     },
 }
